chore: Move test helper functions to `test-fixture`
Fixes #1829
